+++ lib/service_layer/command.rb
@@ -53,6 +53,19 @@ module ServiceLayer
 
     # @!classmethods
     module ClassMethods
+      # @return [Array<Class>] a list of +Exception+.
+      def exceptions
+        @exceptions ||= []
+      end
+
+      # Defines the types of exceptions that must be rescued.
+      #
+      # @param exceptions [Array<Class>] a list of +Exception+.
+      # @return [void]
+      def rescued(*exceptions)
+        self.exceptions.concat exceptions
+      end
+
       # Invokes a +Command+. This is the primary public API method to a command.
       #
       # Delegates the initialization, and performs the service.
@@ -94,7 +107,7 @@ module ServiceLayer
     # @return [Result]
     def execute
       result = perform
-    rescue StandardError => exception
+    rescue *exceptions => exception
       Result.new(error: exception).tap(&:fail!)
     else
       Result.new(result)
@@ -104,6 +117,14 @@ module ServiceLayer
     # @return [Hash]
     def perform
       raise NotImplementedError.new('Service must implement a perform method')
+    end
+
+    private def exceptions
+      if self.class.exceptions.empty?
+        ServiceLayer.configuration.default_exceptions
+      else
+        self.class.exceptions
+      end
     end
   end
 end